requirements.txt
imapclient>=3.1.0
pyyaml>=6.0.2
schedule>=1.2.2
google-api-python-client>=2.164.0
google-auth-oauthlib>=1.2.4
